updates in diar

+++ models/foodConsumerModel.js
@@ -7,11 +7,11 @@ const FoodConsumer = sequelize.define(
     {
         id: { type: DataTypes.INTEGER, autoIncrement: true, primaryKey: true },
         title: { type: DataTypes.STRING, allowNull: false },
-        cals: { type: DataTypes.INTEGER, allowNull: false },
-        carbs: { type: DataTypes.INTEGER, allowNull: false },
-        protein: { type: DataTypes.INTEGER, allowNull: false },
-        fat: { type: DataTypes.INTEGER, allowNull: false },
-        amount: { type: DataTypes.INTEGER, allowNull: false },
+        cals: { type: DataTypes.FLOAT, allowNull: false },
+        carbs: { type: DataTypes.FLOAT, allowNull: false },
+        protein: { type: DataTypes.FLOAT, allowNull: false },
+        fat: { type: DataTypes.FLOAT, allowNull: false },
+        amount: { type: DataTypes.FLOAT, allowNull: false },
         unit: { type: DataTypes.STRING, allowNull: false, defaultValue: 'g' },
     },
     { timestamps: true, createdAt: true, updatedAt: false }
